+++ curation-api/src/main/java/eu/clarin/cmdi/curation/api/report/collection/sec/FileReport.java
@@ -26,6 +26,8 @@ public class FileReport {
    @XmlAttribute
    public double avgScore;
    @XmlElement
+   public String collectionRoot;
+   @XmlElement
    public String provider;
    @XmlElement
    public long numOfFiles;